+++ assets/css/atcb-3d.css
@@ -24,11 +24,13 @@
  */
 
  :root {
-  --base-font-size: 16px;
+  --base-font-size-l: 16px;
+  --base-font-size-m: 16px;
+  --base-font-size-s: 16px;
   --font: arial, helvetica, sans-serif;
   --atcb-keyboard-focus: #1e90ff;
   --atcb-background: #f5f5f5;
-  --atcb-background-2: #fff;
+  --atcb-background-hover: #fff;
   --atcb-border: #d2d2d2;
   --atcb-text: #333;
   --atcb-text-2: #000;
@@ -69,7 +71,7 @@ body.atcb-dark .atcb-modal-box.atcb-bodyScheme,
 #atcb-bgoverlay.atcb-dark,
 body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   --atcb-background: #2e2e2e;
-  --atcb-background-2: #373737;
+  --atcb-background-hover: #373737;
   --atcb-border: #4d4d4d;
   --atcb-text: #dedede;
   --atcb-text-2: #f1f1f1;
@@ -87,13 +89,39 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   --icon-filter: grayscale(.2);
 }
 
+/* Size Breakpoints */
+
+/* large */
+.atcb-button-wrapper,
+.atcb-list,
+.atcb-modal-box {
+  font-size: var(--base-font-size-l);
+}
+
+/* medium */
+@media (max-width: 991px) {
+  .atcb-button-wrapper,
+  .atcb-list,
+  .atcb-modal-box {
+    font-size: var(--base-font-size-m);
+  }
+}
+
+/* small */
+@media (max-width: 575px) {
+  .atcb-button-wrapper,
+  .atcb-list,
+  .atcb-modal-box {
+    font-size: var(--base-font-size-s);
+  }
+}
+
 /******************************
  * The triggering button
  */
 
 .atcb-button-wrapper {
   display: inline-block;
-  font-size: var(--base-font-size);
   padding: 5px;
   position: relative;
 }
@@ -108,7 +136,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   cursor: pointer;
   display: flex;
   font-family: var(--font);
-  font-size: var(--base-font-size);
+  font-size: 1em;
   font-weight: 600;
   justify-content: center;
   line-height: 1.5em;
@@ -134,7 +162,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 
 .atcb-button:focus,
 .atcb-button:hover {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   box-shadow: var(--atcb-shadow-button-hover);
 }
 
@@ -145,7 +173,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 .atcb-button.atcb-active:not(.atcb-modal-style, .atcb-dropoverlay),
 .atcb-button.atcb-single:focus,
 .atcb-button.atcb-single:hover {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   z-index: 15000000;
 }
 
@@ -171,7 +199,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
  */
 
 .atcb-icon {
-  height: var(--base-font-size);
+  height: 1em;
   margin-bottom: .3em;
   margin-right: .8em;
 }
@@ -218,7 +246,6 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   color: var(--atcb-text);
   display: block;
   font-family: var(--font);
-  font-size: var(--base-font-size);
   max-width: 100%;
   position: relative;
   user-select: none;
@@ -295,7 +322,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   box-sizing: border-box;
   cursor: pointer;
   display: flex;
-  font-size: var(--base-font-size);
+  font-size: 1em;
   line-height: 1.75em;
   padding: .8em;
   text-align: left;
@@ -309,12 +336,12 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 }
 
 .atcb-list-item:hover {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   color: var(--atcb-text-2);
 }
 
 .atcb-list-item:focus-visible {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   color: var(--atcb-keyboard-focus);
   font-size: .95em;
   font-weight: 600;
@@ -408,7 +435,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 }
 
 .atcb-modal-box {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   box-shadow: var(--atcb-shadow-modal);
   color: var(--atcb-text-2);
   cursor: default;
@@ -416,7 +443,6 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   border: 1px solid var(--atcb-border);
   box-sizing: border-box;  
   font-family: var(--font);
-  font-size: var(--base-font-size);
   line-height: 1.5em;
   padding: 1.25em 1.5em;
   text-align: left;
@@ -512,7 +538,7 @@ a.atcb-modal-btn {
   cursor: pointer;
   display: inline-block;
   font-family: var(--font);
-  font-size: var(--base-font-size);
+  font-size: 1em;
   font-weight: 600;
   line-height: 1.5em;
   margin: .625em;
@@ -568,7 +594,7 @@ a.atcb-modal-btn:focus-visible {
   min-width: 100%;
   opacity: 0;
   overflow-y: auto;
-  padding: 30px 20px 130px;
+  padding: 20px 20px 130px;
   position: fixed;
   transform: translate3d(0, 0, -10px);
   width: 100vw;
