+++ assets/css/atcb-flat.css
@@ -24,11 +24,13 @@
  */
 
 :root {
-  --base-font-size: 16px;
+  --base-font-size-l: 16px;
+  --base-font-size-m: 16px;
+  --base-font-size-s: 16px;
   --font: arial, helvetica, sans-serif;
   --atcb-keyboard-focus: #1e90ff;
   --atcb-background: #f5f5f5;
-  --atcb-background-2: #fff;
+  --atcb-background-hover: #fff;
   --atcb-border: #545454;
   --atcb-button-shadow: #333;
   --atcb-text: #333;
@@ -63,7 +65,7 @@ body.atcb-dark .atcb-modal-box.atcb-bodyScheme,
 #atcb-bgoverlay.atcb-dark,
 body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   --atcb-background: #222;
-  --atcb-background-2: #373737;
+  --atcb-background-hover: #373737;
   --atcb-border: #515151;
   --atcb-button-shadow: #000;
   --atcb-text: #dedede;
@@ -77,13 +79,39 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   --icon-filter: grayscale(0.2);
 }
 
+/* Size Breakpoints */
+
+/* large */
+.atcb-button-wrapper,
+.atcb-list,
+.atcb-modal-box {
+  font-size: var(--base-font-size-l);
+}
+
+/* medium */
+@media (max-width: 991px) {
+  .atcb-button-wrapper,
+  .atcb-list,
+  .atcb-modal-box {
+    font-size: var(--base-font-size-m);
+  }
+}
+
+/* small */
+@media (max-width: 575px) {
+  .atcb-button-wrapper,
+  .atcb-list,
+  .atcb-modal-box {
+    font-size: var(--base-font-size-s);
+  }
+}
+
 /******************************
  * The triggering button
  */
 
 .atcb-button-wrapper {
   display: inline-block;
-  font-size: var(--base-font-size);
   padding: 5px;
   position: relative;
 }
@@ -96,7 +124,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   cursor: pointer;
   display: flex;
   font-family: var(--font);
-  font-size: var(--base-font-size);
+  font-size: 1em;
   font-weight: 600;
   justify-content: center;
   line-height: 1.5em;
@@ -130,7 +158,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 
 .atcb-button:focus,
 .atcb-button:hover {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   top: 0;
   left: 0;
   box-shadow: 0 0 0 0 var(--atcb-button-shadow);
@@ -143,7 +171,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 .atcb-button.atcb-active:not(.atcb-modal-style, .atcb-dropoverlay),
 .atcb-button.atcb-single:focus,
 .atcb-button.atcb-single:hover {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   top: -5px;
   left: -5px;
   box-shadow: 5px 5px 0 0 var(--atcb-button-shadow);
@@ -159,7 +187,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
  */
 
 .atcb-icon {
-  height: var(--base-font-size);
+  height: 1em;
   margin-right: 1em;
   line-height: 1em;
 }
@@ -216,7 +244,6 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   color: var(--atcb-text);
   display: block;
   font-family: var(--font);
-  font-size: var(--base-font-size);
   max-width: 100%;
   position: relative;
   user-select: none;
@@ -231,7 +258,7 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
   box-sizing: border-box;
   cursor: pointer;
   display: flex;
-  font-size: var(--base-font-size);
+  font-size: 1em;
   line-height: 1.75em;
   padding: 0.8em;
   text-align: left;
@@ -245,12 +272,12 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 }
 
 .atcb-list-item:hover {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   color: var(--atcb-text-2);
 }
 
 .atcb-list-item:focus-visible {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   color: var(--atcb-keyboard-focus);
   font-size: 0.95em;
   font-weight: 600;
@@ -326,13 +353,12 @@ body.atcb-dark #atcb-bgoverlay.atcb-bodyScheme {
 }
 
 .atcb-modal-box {
-  background-color: var(--atcb-background-2);
+  background-color: var(--atcb-background-hover);
   color: var(--atcb-text-2);
   cursor: default;
   border: 1px solid var(--atcb-border);
   box-sizing: border-box;
   font-family: var(--font);
-  font-size: var(--base-font-size);
   line-height: 1.5em;
   padding: 1.25em 1.5em;
   text-align: left;
@@ -480,7 +506,7 @@ a.atcb-modal-btn:focus-visible {
   min-width: 100%;
   opacity: 0;
   overflow-y: auto;
-  padding: 30px 20px 130px;
+  padding: 20px 20px 130px;
   position: fixed;
   width: 100vw;
   z-index: 14000000;
