+++ index.html
@@ -182,8 +182,8 @@
             {
               "name":"Reminder to star the add to calendar button repo",
               "description":"Check out the maybe easiest way to include add to calendar buttons to your website:\n&#8594; [url]https://github.com/add2cal/add-to-calendar-button|Click here![/url]",
-              "startDate":"2022-01-14",
-              "endDate":"2022-01-18",
+              "startDate":"2023-01-14",
+              "endDate":"2023-01-16",
               "location":"World Wide Web",
               "options":[
                 "Apple",
@@ -207,8 +207,8 @@
 {
   "name":"Reminder to star the add to calendar button repo",
   "description":"Check out the maybe easiest way to include add to calendar buttons to your website:\n&#8594; [url]https://github.com/add2cal/add-to-calendar-button|Click here![/url]",
-  "startDate":"2022-01-14",
-  "endDate":"2022-01-18",
+  "startDate":"2023-01-14",
+  "endDate":"2023-01-16",
   "location":"World Wide Web",
   "options":[
     "Apple",
@@ -241,8 +241,8 @@
             {
               "name":"Reminder to star the add to calendar button repo",
               "description":"Check out the maybe easiest way to include add to calendar buttons to your website at:<br>&#8594; [url]https://github.com/add2cal/add-to-calendar-button|Click here![/url]",
-              "startDate":"2022-01-14",
-              "endDate":"2022-01-18",
+              "startDate":"2023-01-14",
+              "endDate":"2023-01-18",
               "location":"World Wide Web",
               "label":"Hit me to save!",
               "options":[
@@ -265,8 +265,8 @@
 {
   "name":"Reminder to star the add to calendar button repo",
   "description":"Check out the maybe easiest way to include add to calendar buttons to your website at:&lt;br&gt;&#8594; [url]https://github.com/add2cal/add-to-calendar-button|Click here![/url]",
-  "startDate":"2022-01-14",
-  "endDate":"2022-01-18",
+  "startDate":"2023-01-14",
+  "endDate":"2023-01-18",
   "location":"World Wide Web",
   "label":"Hit me to save!",
   "options":[
@@ -407,8 +407,8 @@
             {
               "name":"Recurring reminder to star the add to calendar button repo",
               "description":"Check out the maybe easiest way to include add to calendar buttons to your website at:<br>&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
-              "startDate":"2022-07-13",
-              "endDate":"2022-07-13",
+              "startDate":"2023-07-13",
+              "endDate":"2023-07-13",
               "startTime":"10:45",
               "endTime":"15:00",
               "location":"https://github.com/add2cal/add-to-calendar-button",
@@ -431,8 +431,8 @@
 {
   "name":"Recurring reminder to star the add to calendar button repo",
   "description":"Check out the maybe easiest way to include add to calendar buttons to your website at:&lt;br&gt;&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
-  "startDate":"2022-07-13",
-  "endDate":"2022-07-13",
+  "startDate":"2023-07-13",
+  "endDate":"2023-07-13",
   "startTime":"10:45",
   "endTime":"15:00",
   "location":"https://github.com/add2cal/add-to-calendar-button",
@@ -476,16 +476,16 @@
                   "endTime":"23:30"
                 },
                 {
-                  "name":"Event #2/3 to star the add to calendar button repo",
-                  "description":"This is the second part to check the calendar button repository:<br>&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
-                  "startDate":"today+5",
+                  "name":"Event #3/3 to star the add to calendar button repo",
+                  "description":"This is the third part to check the calendar button repository:<br>&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
+                  "startDate":"today+6",
                   "startTime":"13:15",
                   "endTime":"15:05"
                 },
                 {
-                  "name":"Event #3/3 to star the add to calendar button repo",
-                  "description":"This is the third part to check the calendar button repository:<br>&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
-                  "startDate":"today+6",
+                  "name":"Event #2/3 to star the add to calendar button repo",
+                  "description":"This is the second part to check the calendar button repository:<br>&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
+                  "startDate":"today+5",
                   "startTime":"13:15",
                   "endTime":"15:05"
                 }
@@ -498,7 +498,7 @@
                 "Outlook.com",
                 "Yahoo"
               ],
-              "timeZone":"Europe/Berlin",
+              "timeZone":"Europe/London",
               "iCalFileName":"Multi-Date-Event"
             }
           </div>
@@ -557,8 +557,8 @@
             const config = {
               name: "Reminder to star the add to calendar button repo",
               description: "Check out the maybe easiest way to include add to calendar buttons to your website at:<br>&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
-              startDate: "2022-01-14",
-              endDate: "2022-08-18",
+              startDate: "2023-01-14",
+              endDate: "2023-08-18",
               options: ["Google", "iCal"],
               timeZone: "Europe/Berlin",
               trigger: "click",
@@ -576,8 +576,8 @@
   const config = {
     name: "Reminder to star the add to calendar button repo",
     description: "Check out the maybe easiest way to include add to calendar buttons to your website at:&lt;br&gt;&#8594; [url]https://github.com/add2cal/add-to-calendar-button[/url]",
-    startDate: "2022-01-14",
-    endDate: "2022-01-18",
+    startDate: "2023-01-14",
+    endDate: "2023-01-18",
     options: ["Google", "iCal"],
     timeZone: "Europe/Berlin",
     trigger: "click",
@@ -616,8 +616,8 @@
                 e.preventDefault();
                 atcb_action({
                   name: e.currentTarget.elements.name.value,
-                  startDate: "2022-01-14",
-                  endDate: "2022-01-18",
+                  startDate: "2023-01-14",
+                  endDate: "2023-01-18",
                   options: ["Apple", "Google", "iCal", "Microsoft365", "Outlook.com", "MicrosoftTeams", "Yahoo"],
                   timeZone: "Europe/Berlin",
                   listStyle: "modal",
@@ -643,8 +643,8 @@
       e.preventDefault();
       atcb_action({
         name: e.currentTarget.elements.name.value,
-        startDate: "2022-01-14",
-        endDate: "2022-01-18",
+        startDate: "2023-01-14",
+        endDate: "2023-01-18",
         options: ["Apple", "Google", "iCal", "Microsoft365", "Outlook.com", "MicrosoftTeams", "Yahoo"],
         timeZone: "Europe/Berlin",
         listStyle: "modal",
